+++ com.tencent.mobileqq/classes.jar/com/tencent/mobileqq/app/BrowserAppInterface$4.java
@@ -15,21 +15,19 @@ class BrowserAppInterface$4
       if (QLog.isColorLevel()) {
         QLog.d("BrowserAppInterface", 2, "no activity running, reboot for tbs now");
       }
-      localIntent = new Intent();
+      Intent localIntent = new Intent();
       localIntent.putExtra("qq_mode_foreground", true);
       BrowserAppInterface.a(this.this$0, localIntent);
-    }
-    while (!QLog.isColorLevel())
-    {
-      Intent localIntent;
       return;
     }
-    QLog.d("BrowserAppInterface", 2, "activity still running, cannot reboot");
+    if (QLog.isColorLevel()) {
+      QLog.d("BrowserAppInterface", 2, "activity still running, cannot reboot");
+    }
   }
 }
 
 
-/* Location:           L:\local\mybackup\temp\qq_apk\com.tencent.mobileqq\classes7.jar
+/* Location:           L:\local\mybackup\temp\qq_apk\com.tencent.mobileqq\classes5.jar
  * Qualified Name:     com.tencent.mobileqq.app.BrowserAppInterface.4
  * JD-Core Version:    0.7.0.1
  */
